+++ packages/gladia/src/gladia-transcription-model.ts
@@ -7,6 +7,7 @@ import {
   combineHeaders,
   convertBase64ToUint8Array,
   createJsonResponseHandler,
+  mediaTypeToExtension,
   delay,
   getFromApi,
   parseProviderOptions,
@@ -497,10 +498,11 @@ export class GladiaTranscriptionModel implements TranscriptionModelV2 {
         ? new Blob([options.audio])
         : new Blob([convertBase64ToUint8Array(options.audio)]);
 
-    formData.append('model', this.modelId);
+    const fileExtension = mediaTypeToExtension(options.mediaType);
     formData.append(
       'audio',
       new File([blob], 'audio', { type: options.mediaType }),
+      `audio.${fileExtension}`,
     );
 
     const { value: uploadResponse } = await postFormDataToApi({